Activate S104: Files should not have too many lines of code
