+++ src/Management/src/Endpoint/Actuators/HeapDump/DotNetHeapDump/Graph.cs
@@ -663,448 +663,6 @@ namespace Graphs
     }
 
     /// <summary>
-    /// Node represents a single node in the code:Graph.  These are created lazily and follow a pattern were the 
-    /// CALLER provides the storage for any code:Node or code:NodeType value that are returned.   Thus the caller
-    /// is responsible for determine when nodes can be reused to minimize GC cost.  
-    /// 
-    /// A node implicitly knows where the 'next' child is (that is it is an iterator).  
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    internal class Node
-    {
-        public int Size
-        {
-            get
-            {
-                m_graph.m_reader.Goto(m_graph.m_nodes[(int)m_index]);
-                var typeAndSize = ReadCompressedInt(m_graph.m_reader);
-                if ((typeAndSize & 1) != 0)     // low bit indicates if Size is encoded explicitly
-                {
-                    return ReadCompressedInt(m_graph.m_reader);
-                }
-
-                // Then it is in the type;
-                typeAndSize >>= 1;
-                return m_graph.m_types[typeAndSize].Size;
-            }
-        }
-        public bool Defined { get { return m_graph.IsDefined(Index); } }
-        public NodeType GetType(NodeType storage)
-        {
-            return m_graph.GetType(TypeIndex, storage);
-        }
-
-        /// <summary>
-        /// Reset the internal state so that 'GetNextChildIndex; will return the first child.  
-        /// </summary>
-        public void ResetChildrenEnumeration()
-        {
-            m_graph.m_reader.Goto(m_graph.m_nodes[(int)m_index]);
-            if ((ReadCompressedInt(m_graph.m_reader) & 1) != 0)        // Skip nodeId and Size
-            {
-                ReadCompressedInt(m_graph.m_reader);
-            }
-
-            m_numChildrenLeft = ReadCompressedInt(m_graph.m_reader);
-            Debug.Assert(m_numChildrenLeft < 1660000);     // Not true in general but good enough for unit testing.
-            m_current = m_graph.m_reader.Current;
-        }
-
-        /// <summary>
-        /// Gets the index of the first child of node.  Will return NodeIndex.Invalid if there are no children. 
-        /// </summary>
-        /// <returns>The index of the child </returns>
-        public NodeIndex GetFirstChildIndex()
-        {
-            ResetChildrenEnumeration();
-            return GetNextChildIndex();
-        }
-        public NodeIndex GetNextChildIndex()
-        {
-            if (m_numChildrenLeft == 0)
-            {
-                return NodeIndex.Invalid;
-            }
-
-            m_graph.m_reader.Goto(m_current);
-
-            var ret = (NodeIndex)(ReadCompressedInt(m_graph.m_reader) + (int)m_index);
-            Debug.Assert((uint)ret < (uint)m_graph.NodeIndexLimit);
-
-            m_current = m_graph.m_reader.Current;
-            --m_numChildrenLeft;
-            return ret;
-        }
-
-        /// <summary>
-        /// Returns the number of children this node has.  
-        /// </summary>
-        public int ChildCount
-        {
-            get
-            {
-                m_graph.m_reader.Goto(m_graph.m_nodes[(int)m_index]);
-                if ((ReadCompressedInt(m_graph.m_reader) & 1) != 0)        // Skip nodeId and Size
-                {
-                    ReadCompressedInt(m_graph.m_reader);
-                }
-
-                return ReadCompressedInt(m_graph.m_reader);
-            }
-        }
-        public NodeTypeIndex TypeIndex
-        {
-            get
-            {
-                m_graph.m_reader.Goto(m_graph.m_nodes[(int)m_index]);
-                var ret = (NodeTypeIndex)(ReadCompressedInt(m_graph.m_reader) >> 1);
-                return ret;
-            }
-        }
-        public NodeIndex Index { get { return m_index; } }
-        public Graph Graph { get { return m_graph; } }
-        /// <summary>
-        /// Returns true if 'node' is a child of 'this'.  childStorage is simply used as temp space 
-        /// as was allocated by Graph.AllocateNodeStorage
-        /// </summary>
-        public bool Contains(NodeIndex nodeIndex)
-        {
-            for (NodeIndex childIndex = GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = GetNextChildIndex())
-            {
-                if (childIndex == nodeIndex)
-                {
-                    return true;
-                }
-            }
-            return false;
-        }
-
-        public override string ToString()
-        {
-            StringWriter sw = new StringWriter();
-            WriteXml(sw, includeChildren: false);
-            return sw.ToString();
-        }
-        public virtual void WriteXml(TextWriter writer, bool includeChildren = true, string prefix = "", NodeType typeStorage = null, string additinalAttribs = "")
-        {
-            Debug.Assert(Index != NodeIndex.Invalid);
-            typeStorage ??= m_graph.AllocTypeNodeStorage();
-
-            if (m_graph.m_nodes[(int)Index] == StreamLabel.Invalid)
-            {
-                writer.WriteLine("{0}<Node Index=\"{1}\" Undefined=\"true\"{2}/>", prefix, (int)Index, additinalAttribs);
-                return;
-            }
-
-            writer.Write("{0}<Node Index=\"{1}\" TypeIndex=\"{2}\" Size=\"{3}\" Type=\"{4}\" NumChildren=\"{5}\"{6}",
-                prefix, (int)Index, TypeIndex, Size, SecurityElement.Escape(GetType(typeStorage).Name),
-                ChildCount, additinalAttribs);
-            var childIndex = GetFirstChildIndex();
-            if (childIndex != NodeIndex.Invalid)
-            {
-                writer.WriteLine(">");
-                if (includeChildren)
-                {
-                    writer.Write(prefix);
-                    int i = 0;
-                    do
-                    {
-                        writer.Write(" {0}", childIndex);
-                        childIndex = GetNextChildIndex();
-                        i++;
-                        if (i >= 32)
-                        {
-                            writer.WriteLine();
-                            writer.Write(prefix);
-                            i = 0;
-                        }
-                    } while (childIndex != NodeIndex.Invalid);
-                }
-                else
-                {
-                    writer.Write(prefix);
-                    writer.WriteLine($"<!-- {ChildCount} children omitted... -->");
-                }
-                writer.WriteLine(" </Node>");
-            }
-            else
-            {
-                writer.WriteLine("/>");
-            }
-        }
-        #region private
-        protected internal Node(Graph graph)
-        {
-            m_graph = graph;
-            m_index = NodeIndex.Invalid;
-        }
-
-        // Node information is stored in a compressed form because we have a lot of them. 
-        internal static int ReadCompressedInt(SegmentedMemoryStreamReader reader)
-        {
-            int ret = 0;
-            byte b = reader.ReadByte();
-            ret = b << 25 >> 25;
-            if ((b & 0x80) == 0)
-            {
-                return ret;
-            }
-
-            ret <<= 7;
-            b = reader.ReadByte();
-            ret += b & 0x7f;
-            if ((b & 0x80) == 0)
-            {
-                return ret;
-            }
-
-            ret <<= 7;
-            b = reader.ReadByte();
-            ret += b & 0x7f;
-            if ((b & 0x80) == 0)
-            {
-                return ret;
-            }
-
-            ret <<= 7;
-            b = reader.ReadByte();
-            ret += b & 0x7f;
-            if ((b & 0x80) == 0)
-            {
-                return ret;
-            }
-
-            ret <<= 7;
-            b = reader.ReadByte();
-            Debug.Assert((b & 0x80) == 0);
-            ret += b;
-            return ret;
-        }
-
-        internal static void WriteCompressedInt(SegmentedMemoryStreamWriter writer, int value)
-        {
-            if (value << 25 >> 25 == value)
-            {
-                goto oneByte;
-            }
-
-            if (value << 18 >> 18 == value)
-            {
-                goto twoBytes;
-            }
-
-            if (value << 11 >> 11 == value)
-            {
-                goto threeBytes;
-            }
-
-            if (value << 4 >> 4 == value)
-            {
-                goto fourBytes;
-            }
-
-            writer.Write((byte)((value >> 28) | 0x80));
-            fourBytes:
-            writer.Write((byte)((value >> 21) | 0x80));
-            threeBytes:
-            writer.Write((byte)((value >> 14) | 0x80));
-            twoBytes:
-            writer.Write((byte)((value >> 7) | 0x80));
-            oneByte:
-            writer.Write((byte)(value & 0x7F));
-        }
-
-        internal NodeIndex m_index;
-        internal Graph m_graph;
-        private StreamLabel m_current;          // My current child in the enumerable.
-        private int m_numChildrenLeft;          // count of my children
-        #endregion
-    }
-
-    /// <summary>
-    /// Represents the nodeId of a particular node in the graph.  
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    internal sealed class NodeType
-    {
-        /// <summary>
-        /// Every nodeId has a name, this is it.  
-        /// </summary>
-        public string Name
-        {
-            get
-            {
-                var ret = m_graph.m_types[(int)m_index].Name;
-                if (ret == null && (int)m_index < m_graph.m_deferedTypes.Count)
-                {
-                    var info = m_graph.m_deferedTypes[(int)m_index];
-                    if (m_graph.ResolveTypeName != null)
-                    {
-                        ret = m_graph.ResolveTypeName(info.TypeID, info.Module);
-                        if (info.TypeNameSuffix != null)
-                        {
-                            ret += info.TypeNameSuffix;
-                        }
-
-                        m_graph.m_types.UnderlyingArray[(int)m_index].Name = ret;
-                    }
-                    ret ??= $"TypeID(0x{info.TypeID:x})";
-                }
-                return ret;
-            }
-        }
-        /// <summary>
-        /// This is the ModuleName ! Name (or just Name if ModuleName does not exist)  
-        /// </summary>
-        public string FullName
-        {
-            get
-            {
-                var moduleName = ModuleName;
-                if (moduleName == null)
-                {
-                    return Name;
-                }
-
-                if (moduleName.Length == 0) // TODO should we have this convention?   
-                {
-                    moduleName = "?";
-                }
-
-                return $"{moduleName}!{Name}";
-            }
-        }
-        /// <summary>
-        /// Size is defined as the Size of the first node in the graph of a given nodeId.   
-        /// For types that always have the same Size this is useful, but for types (like arrays or strings)
-        /// that have variable Size, it is not useful.  
-        /// 
-        /// TODO keep track if the nodeId is of variable Size
-        /// </summary>
-        public int Size { get { return m_graph.m_types[(int)m_index].Size; } }
-        public NodeTypeIndex Index { get { return m_index; } }
-        public Graph Graph { get { return m_graph; } }
-        /// <summary>
-        /// The module associated with the type.  Can be null.  Typically this is the full path name.  
-        /// </summary>
-        public string ModuleName
-        {
-            get
-            {
-                var ret = m_graph.m_types[(int)m_index].ModuleName;
-                if (ret == null && (int)m_index < m_graph.m_deferedTypes.Count)
-                {
-                    var module = m_graph.m_deferedTypes[(int)m_index].Module;
-                    if (module != null)
-                    {
-                        ret = module.Path;
-                    }
-                }
-                return ret;
-            }
-            set
-            {
-                var typeInfo = m_graph.m_types[(int)m_index];
-                typeInfo.ModuleName = value;
-                m_graph.m_types[(int)m_index] = typeInfo;
-            }
-        }
-        public Module Module { get { return m_graph.m_deferedTypes[(int)m_index].Module; } }
-        public int RawTypeID { get { return m_graph.m_deferedTypes[(int)m_index].TypeID; } }
-
-        public override string ToString()
-        {
-            StringWriter sw = new StringWriter();
-            WriteXml(sw);
-            return sw.ToString();
-        }
-        public void WriteXml(TextWriter writer, string prefix = "")
-        {
-            writer.WriteLine("{0}<NodeType Index=\"{1}\" Name=\"{2}\"/>", prefix, (int)Index, SecurityElement.Escape(Name));
-        }
-        #region private
-        protected internal NodeType(Graph graph)
-        {
-            m_graph = graph;
-            m_index = NodeTypeIndex.Invalid;
-        }
-
-        internal Graph m_graph;
-        internal NodeTypeIndex m_index;
-        #endregion
-    }
-
-    /// <summary>
-    /// Holds all interesting data about a module (in particular enough to look up PDB information)
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    public class Module : IFastSerializable
-    {
-        /// <summary>
-        /// Create new module.  You must have at least a image base.   Everything else is optional.
-        /// </summary>
-        public Module(Address imageBase) { ImageBase = imageBase; }
-
-        /// <summary>
-        /// The path to the Module (can be null if not known)
-        /// </summary>
-        public string Path;
-        /// <summary>
-        /// The location where the image was loaded into memory
-        /// </summary>
-        public Address ImageBase;
-        /// <summary>
-        /// The size of the image when loaded in memory
-        /// </summary>
-        public int Size;
-        /// <summary>
-        /// The time when this image was built (There is a field in the PE header).   May be MinimumValue if unknown. 
-        /// </summary>
-        public DateTime BuildTime;      // From in the PE header
-        /// <summary>
-        /// The name of hte PDB file associated with this module.   Ma bye null if unknown
-        /// </summary>
-        public string PdbName;
-        /// <summary>
-        /// The GUID that uniquely identifies this PDB for symbol server lookup.  May be Guid.Empty it not known.  
-        /// </summary>
-        public Guid PdbGuid;            // PDB Guid 
-        /// <summary>
-        /// The age (version number) that is used for symbol server lookup.  
-        /// </summary>T
-        public int PdbAge;
-
-        #region private
-        /// <summary>
-        /// Implementing IFastSerializable interface.  
-        /// </summary>
-        public void ToStream(Serializer serializer)
-        {
-            serializer.Write(Path);
-            serializer.Write((long)ImageBase);
-            serializer.Write(Size);
-            serializer.Write(BuildTime.Ticks);
-            serializer.Write(PdbName);
-            serializer.Write(PdbGuid);
-            serializer.Write(PdbAge);
-        }
-        /// <summary>
-        /// Implementing IFastSerializable interface.  
-        /// </summary>
-        public void FromStream(Deserializer deserializer)
-        {
-            deserializer.Read(out Path);
-            ImageBase = (Address)deserializer.ReadInt64();
-            deserializer.Read(out Size);
-            BuildTime = new DateTime(deserializer.ReadInt64());
-            deserializer.Read(out PdbName);
-            deserializer.Read(out PdbGuid);
-            deserializer.Read(out PdbAge);
-        }
-        #endregion
-    }
-
-    /// <summary>
     /// Each node is given a unique index (which is dense: an array is a good lookup structure).   
     /// To avoid passing the wrong indexes to methods, we make an enum for each index.   This does
     /// mean you need to cast away this strong typing occasionally (e.g. when you index arrays)
@@ -1118,1535 +676,6 @@ namespace Graphs
     /// However on the whole it is a good tradeoff.  
     /// </summary>    
     public enum NodeTypeIndex { Invalid = -1 }
-
-    /// <summary>
-    /// Stuff that is useful but does not need to be in Graph.   
-    /// </summary>
-    [ExcludeFromCodeCoverage]
-    internal static class GraphUtils
-    {
-        /// <summary>
-        /// Write the graph as XML to a string and return it (useful for debugging small graphs).  
-        /// </summary>
-        /// <returns></returns>
-        public static string PrintGraph(this Graph graph)
-        {
-            StringWriter sw = new StringWriter();
-            graph.WriteXml(sw);
-            return sw.ToString();
-        }
-        public static string PrintNode(this Graph graph, NodeIndex nodeIndex)
-        {
-            return graph.GetNode(nodeIndex, graph.AllocNodeStorage()).ToString();
-        }
-        public static string PrintNode(this Graph graph, int nodeIndex)
-        {
-            return graph.PrintNode((NodeIndex)nodeIndex);
-        }
-        public static string PrintNodes(this Graph graph, List<NodeIndex> nodes)
-        {
-            var sw = new StringWriter();
-            sw.WriteLine("<NodeList>");
-            var node = graph.AllocNodeStorage();
-            var type1 = graph.AllocTypeNodeStorage();
-
-            foreach (var nodeIndex in nodes)
-            {
-                node = graph.GetNode(nodeIndex, node);
-                node.WriteXml(sw, prefix: "  ", typeStorage: type1);
-            }
-            sw.WriteLine("<NodeList>");
-            return sw.ToString();
-        }
-        public static string PrintChildren(this Graph graph, NodeIndex nodeIndex)
-        {
-            return graph.PrintNodes(graph.NodeChildren(nodeIndex));
-        }
-        public static string PrintChildren(this Graph graph, int nodeIndex)
-        {
-            return graph.PrintChildren((NodeIndex)nodeIndex);
-        }
-        // Debugging. 
-        /// <summary>
-        /// Writes the graph as XML to 'writer'.  Don't use on big graphs.  
-        /// </summary>
-        public static void WriteXml(this Graph graph, TextWriter writer)
-        {
-            writer.WriteLine("<MemoryGraph NumNodes=\"{0}\" NumTypes=\"{1}\" TotalSize=\"{2}\" SizeOfGraphDescription=\"{3}\">",
-                graph.NodeIndexLimit, graph.NodeTypeIndexLimit, graph.TotalSize, graph.SizeOfGraphDescription());
-            writer.WriteLine(" <RootIndex>{0}</RootIndex>", graph.RootIndex);
-            writer.WriteLine(" <NodeTypes Count=\"{0}\">", graph.NodeTypeIndexLimit);
-            var typeStorage = graph.AllocTypeNodeStorage();
-            for (NodeTypeIndex typeIndex = 0; typeIndex < graph.NodeTypeIndexLimit; typeIndex++)
-            {
-                var type = graph.GetType(typeIndex, typeStorage);
-                type.WriteXml(writer, "  ");
-            }
-            writer.WriteLine(" </NodeTypes>");
-
-            writer.WriteLine(" <Nodes Count=\"{0}\">", graph.NodeIndexLimit);
-            var nodeStorage = graph.AllocNodeStorage();
-            for (NodeIndex nodeIndex = 0; nodeIndex < graph.NodeIndexLimit; nodeIndex++)
-            {
-                var node = graph.GetNode(nodeIndex, nodeStorage);
-                node.WriteXml(writer, prefix: "  ");
-            }
-            writer.WriteLine(" </Nodes>");
-            writer.WriteLine("</MemoryGraph>");
-        }
-        public static void DumpNormalized(this MemoryGraph graph, TextWriter writer)
-        {
-            MemoryNode nodeStorage = (MemoryNode)graph.AllocNodeStorage();
-            NodeType typeStorage = graph.AllocTypeNodeStorage();
-            Node node;
-
-#if false 
-            // Compute reachability info
-            bool[] reachable = new bool[(int)graph.NodeIndexLimit];
-            Queue<NodeIndex> workQueue = new Queue<NodeIndex>();
-            workQueue.Enqueue(graph.RootIndex);
-            while (workQueue.Count > 0)
-            {
-                var nodeIdx = workQueue.Dequeue();
-                if (!reachable[(int)nodeIdx])
-                {
-                    reachable[(int)nodeIdx] = true;
-                    node = graph.GetNode(nodeIdx, nodeStorage);
-                    for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-                        workQueue.Enqueue(childIndex);
-                }
-            }
-
-            // Get Reachability count. 
-            int reachableCount = 0;
-            for (int i = 0; i < reachable.Length; i++)
-                if (reachable[i])
-                    reachableCount++;
-#endif
-
-            // Sort the nodes by virtual address 
-            NodeIndex[] sortedNodes = new NodeIndex[(int)graph.NodeIndexLimit];
-            for (int i = 0; i < sortedNodes.Length; i++)
-            {
-                sortedNodes[i] = (NodeIndex)i;
-            }
-
-            Array.Sort(sortedNodes, delegate (NodeIndex x, NodeIndex y)
-            {
-                // Sort first by address
-                int ret = graph.GetAddress(x).CompareTo(graph.GetAddress(y));
-                if (ret != 0)
-                {
-                    return ret;
-                }
-                // Then by name
-                return graph.GetNode(x, nodeStorage).GetType(typeStorage).Name.CompareTo(graph.GetNode(y, nodeStorage).GetType(typeStorage).Name);
-            });
-
-            node = graph.GetNode(graph.RootIndex, nodeStorage);
-            writer.WriteLine("<GraphDump RootNode=\"{0}\" NumNodes=\"{1}\" NumTypes=\"{2}\" TotalSize=\"{3}\" SizeOfGraphDescription=\"{4}\">",
-                SecurityElement.Escape(node.GetType(typeStorage).Name),
-                graph.NodeIndexLimit,
-                graph.NodeTypeIndexLimit,
-                graph.TotalSize,
-                graph.SizeOfGraphDescription());
-            writer.WriteLine(" <Nodes Count=\"{0}\">", graph.NodeIndexLimit);
-
-            SortedDictionary<ulong, bool> roots = new SortedDictionary<ulong, bool>();
-            foreach (NodeIndex nodeIdx in sortedNodes)
-            {
-                // if (!reachable[(int)nodeIdx]) continue;
-
-                node = graph.GetNode(nodeIdx, nodeStorage);
-                string name = node.GetType(typeStorage).Name;
-
-                writer.Write("  <Node Address=\"{0:x}\" Size=\"{1}\" Type=\"{2}\"> ", graph.GetAddress(nodeIdx), node.Size, SecurityElement.Escape(name));
-                bool isRoot = graph.GetAddress(node.Index) == 0;
-                int childCnt = 0;
-                for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-                {
-                    if (isRoot)
-                    {
-                        roots[graph.GetAddress(childIndex)] = true;
-                    }
-
-                    childCnt++;
-                    if (childCnt % 8 == 0)
-                    {
-                        writer.WriteLine();
-                        writer.Write("    ");
-                    }
-                    writer.Write("{0:x} ", graph.GetAddress(childIndex));
-                }
-                writer.WriteLine(" </Node>");
-            }
-            writer.WriteLine(" <Roots>");
-            foreach (ulong root in roots.Keys)
-            {
-                writer.WriteLine("  {0:x}", root);
-            }
-            writer.WriteLine(" </Roots>");
-            writer.WriteLine(" </Nodes>");
-            writer.WriteLine("</GraphDump>");
-        }
-
-        public static List<NodeIndex> NodeChildren(this Graph graph, NodeIndex nodeIndex)
-        {
-            var node = graph.GetNode(nodeIndex, graph.AllocNodeStorage());
-            var ret = new List<NodeIndex>();
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                ret.Add(childIndex);
-            }
-
-            return ret;
-        }
-        public static List<NodeIndex> NodesOfType(this Graph graph, string regExpression)
-        {
-            var typeSet = new Dictionary<NodeTypeIndex, NodeTypeIndex>();
-            var type = graph.AllocTypeNodeStorage();
-            for (NodeTypeIndex typeId = 0; typeId < graph.NodeTypeIndexLimit; typeId = typeId + 1)
-            {
-                type = graph.GetType(typeId, type);
-                if (Regex.IsMatch(type.Name, regExpression))
-                {
-                    typeSet.Add(typeId, typeId);
-                }
-            }
-
-            var ret = new List<NodeIndex>();
-            var node = graph.AllocNodeStorage();
-            for (NodeIndex nodeId = 0; nodeId < graph.NodeIndexLimit; nodeId = nodeId + 1)
-            {
-                node = graph.GetNode(nodeId, node);
-                if (typeSet.ContainsKey(node.TypeIndex))
-                {
-                    ret.Add(nodeId);
-                }
-            }
-            return ret;
-        }
-    }
-}
-
-/// <summary>
-/// A RefGraph is derived graph where each node's children are the set of nodes in the original graph 
-/// which refer that node (that is A -> B then in refGraph B -> A).   
-/// 
-/// The NodeIndexes in the refGraph match the NodeIndexes in the original graph.  Thus after creating
-/// a refGraph it is easy to answer the question 'who points at me' of the original graph.  
-/// 
-/// When create the RefGraph the whole reference graph is generated on the spot (thus it must traverse
-/// the whole of the original graph) and the size of the resulting RefGraph is  about the same size as the  
-/// original graph. 
-/// 
-/// Thus this is a fairly expensive thing to create.  
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class RefGraph
-{
-    public RefGraph(Graph graph)
-    {
-        m_refsForNodes = new NodeListIndex[(int)graph.NodeIndexLimit];
-        // We guess that we need about 1.5X as many slots as there are nodes.   This seems a conservative estimate. 
-        m_links = new GrowableArray<RefElem>((int)graph.NodeIndexLimit * 3 / 2);
-
-        var nodeStorage = graph.AllocNodeStorage();
-        for (NodeIndex nodeIndex = 0; nodeIndex < graph.NodeIndexLimit; nodeIndex++)
-        {
-            var node = graph.GetNode(nodeIndex, nodeStorage);
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                AddRefsTo(childIndex, nodeIndex);
-            }
-        }
-
-        // Sadly, this check is too expensive even for DEBUG 
-#if false 
-        CheckConsistancy(graph);
-#endif
-    }
-    /// <summary>
-    /// Allocates nodes to be used as storage for methods like code:GetNode, code:RefNode.GetFirstChild and code:RefNode.GetNextChild
-    /// </summary>
-    public RefNode AllocNodeStorage() { return new RefNode(this); }
-
-    /// <summary>
-    /// Given an arbitrary code:NodeIndex that identifies the node, Get a code:Node object.  
-    /// 
-    /// This routine does not allocated but uses the space passed in by 'storage.  
-    /// 'storage' should be allocated with code:AllocNodeStorage, and should be aggressively reused.  
-    /// </summary>
-    public RefNode GetNode(NodeIndex nodeIndex, RefNode storage)
-    {
-        Debug.Assert(storage.m_graph == this);
-        storage.m_index = nodeIndex;
-        return storage;
-    }
-
-    /// <summary>
-    /// This is for debugging 
-    /// </summary>
-    /// <param name="nodeIndex"></param>
-    /// <returns></returns>
-    public RefNode GetNode(NodeIndex nodeIndex)
-    {
-        return GetNode(nodeIndex, AllocNodeStorage());
-    }
-
-    #region private
-#if DEBUG
-    private void CheckConsitancy(Graph graph)
-    {
-        // This double check is pretty expensive for large graphs (nodes that have large fan-in or fan-out).  
-        var nodeStorage = graph.AllocNodeStorage();
-        var refStorage = AllocNodeStorage();
-        for (NodeIndex nodeIdx = 0; nodeIdx < graph.NodeIndexLimit; nodeIdx++)
-        {
-            // If Node -> Ref then the RefGraph has a pointer from Ref -> Node 
-            var node = graph.GetNode(nodeIdx, nodeStorage);
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                var refsForChild = GetNode(childIndex, refStorage);
-                if (!refsForChild.Contains(nodeIdx))
-                {
-                    _ = node.ToString();
-                    _ = refsForChild.ToString();
-                    Debug.Assert(false);
-                }
-            }
-
-            // If the refs graph has a pointer from Ref -> Node then the original graph has a arc from Node ->Ref
-            var refNode = GetNode(nodeIdx, refStorage);
-            for (var childIndex = refNode.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = refNode.GetNextChildIndex())
-            {
-                var nodeForChild = graph.GetNode(childIndex, nodeStorage);
-                if (!nodeForChild.Contains(nodeIdx))
-                {
-                    _ = nodeForChild.ToString();
-                    _ = refNode.ToString();
-                    Debug.Assert(false);
-                }
-            }
-        }
-    }
-#endif
-
-    /// <summary>
-    /// Add the fact that 'refSource' refers to refTarget.
-    /// </summary>
-    private void AddRefsTo(NodeIndex refTarget, NodeIndex refSource)
-    {
-        NodeListIndex refsToList = m_refsForNodes[(int)refTarget];
-
-        // We represent singles as the childIndex itself.  This is a very common case, so it is good that it is efficient. 
-        if (refsToList == NodeListIndex.Empty)
-        {
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(refSource + 1);
-        }
-        else if (refsToList > 0)        // One element list
-        {
-            var existingChild = (NodeIndex)(refsToList - 1);
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(-AddLink(refSource, AddLink(existingChild)) - 1);
-        }
-        else // refsToList < 0          more than one element.  
-        {
-            var listIndex = -(int)refsToList - 1;
-            m_refsForNodes[(int)refTarget] = (NodeListIndex)(-AddLink(refSource, listIndex) - 1);
-        }
-    }
-
-    /// <summary>
-    /// A helper function for AddRefsTo.  Allocates a new cell from m_links and initializes its two fields 
-    /// (the child index field and 'rest' field), and returns the index (pointer) to the new cell.  
-    /// </summary>
-    private int AddLink(NodeIndex refIdx, int nextIdx = -1)
-    {
-        var ret = m_links.Count;
-        m_links.Add(new RefElem(refIdx, nextIdx));
-        return ret;
-    }
-
-    /// <summary>
-    ///  Logically a NodeListIndex represents a list of node indexes.   However it is heavily optimized
-    ///  to avoid overhead.   0 means empty, a positive number is the NodeIndex+1 and a negative number 
-    ///  is index in m_links - 1.  
-    /// </summary>
-    internal enum NodeListIndex { Empty = 0 };
-
-    /// <summary>
-    /// RefElem is a linked list cell that is used to store lists of children that are larger than 1.
-    /// </summary>
-    internal struct RefElem
-    {
-        public RefElem(NodeIndex refIdx, int nextIdx) { RefIdx = refIdx; NextIdx = nextIdx; }
-        public NodeIndex RefIdx;           // The reference
-        public int NextIdx;                // The index to the next element in  m_links.   a negative number when done. 
-    }
-
-    /// <summary>
-    /// m_refsForNodes maps the NodeIndexes of the reference graph to the children information.   However unlike
-    /// a normal Graph RefGraph needs to support incremental addition of children.  Thus we can't use the normal
-    /// compression (which assumed you know all the children when you define the node).  
-    /// 
-    /// m_refsForNodes points at a NodeListIndex which is a compressed list that is tuned for the case where
-    /// a node has exactly one child (a very common case).   If that is not true we 'overflow' into a 'linked list'
-    /// of RefElems that is stored in m_links.   See NodeListIndex for more on the exact encoding.   
-    /// 
-    /// </summary>
-    internal NodeListIndex[] m_refsForNodes;
-
-    /// <summary>
-    /// If the number of children for a node is > 1 then we need to store the data somewhere.  m_links is array
-    /// of linked list cells that hold the overflow case.  
-    /// </summary>
-    internal GrowableArray<RefElem> m_links;      // The rest of the list.  
-    #endregion
-}
-
-[ExcludeFromCodeCoverage]
-internal sealed class RefNode
-{
-    /// <summary>
-    /// Gets the first child for the node.  Will return null if there are no children.  
-    /// </summary>
-    public NodeIndex GetFirstChildIndex()
-    {
-        var refsToList = m_graph.m_refsForNodes[(int)m_index];
-
-        if (refsToList == RefGraph.NodeListIndex.Empty)
-        {
-            return NodeIndex.Invalid;
-        }
-
-        if (refsToList > 0)        // One element list
-        {
-            m_cur = -1;
-            return (NodeIndex)(refsToList - 1);
-        }
-        else // refsToList < 0          more than one element.  
-        {
-            var listIndex = -(int)refsToList - 1;
-            var refElem = m_graph.m_links[listIndex];
-            m_cur = refElem.NextIdx;
-            return refElem.RefIdx;
-        }
-    }
-    /// <summary>
-    /// Returns the next child for the node.   Will return NodeIndex.Invalid if there are no more children 
-    /// </summary>
-    public NodeIndex GetNextChildIndex()
-    {
-        if (m_cur < 0)
-        {
-            return NodeIndex.Invalid;
-        }
-
-        var refElem = m_graph.m_links[m_cur];
-        m_cur = refElem.NextIdx;
-        return refElem.RefIdx;
-    }
-
-    /// <summary>
-    /// Returns the count of children (nodes that reference this node). 
-    /// </summary>
-    public int ChildCount
-    {
-        get
-        {
-            var ret = 0;
-            for (NodeIndex childIndex = GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = GetNextChildIndex())
-            {
-                ret++;
-            }
-
-            return ret;
-        }
-    }
-
-    public RefGraph Graph { get { return m_graph; } }
-    public NodeIndex Index { get { return m_index; } }
-
-    /// <summary>
-    /// Returns true if 'node' is a child of 'this'.  childStorage is simply used as temp space 
-    /// as was allocated by RefGraph.AllocateNodeStorage
-    /// </summary>
-    public bool Contains(NodeIndex node)
-    {
-        for (NodeIndex childIndex = GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = GetNextChildIndex())
-        {
-            if (childIndex == node)
-            {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public override string ToString()
-    {
-        StringWriter sw = new StringWriter();
-        WriteXml(sw);
-        return sw.ToString();
-    }
-    public void WriteXml(TextWriter writer, string prefix = "")
-    {
-        Debug.Assert(Index != NodeIndex.Invalid);
-
-
-        writer.Write("{0}<Node Index=\"{1}\" NumChildren=\"{2}\"", prefix, (int)Index, ChildCount);
-        var childIndex = GetFirstChildIndex();
-        if (childIndex != NodeIndex.Invalid)
-        {
-            writer.WriteLine(">");
-            writer.Write(prefix);
-            int i = 0;
-            do
-            {
-                writer.Write(" {0}", childIndex);
-                childIndex = GetNextChildIndex();
-                i++;
-                if (i >= 32)
-                {
-                    writer.WriteLine();
-                    writer.Write(prefix);
-                    i = 0;
-                }
-            } while (childIndex != NodeIndex.Invalid);
-            writer.WriteLine(" </Node>");
-        }
-        else
-        {
-            writer.WriteLine("/>");
-        }
-    }
-
-    #region private
-    internal RefNode(RefGraph refGraph)
-    {
-        m_graph = refGraph;
-    }
-
-    internal RefGraph m_graph;
-    internal NodeIndex m_index;     // My index.  
-    internal int m_cur;             // A pointer to where we are in the list of elements (index into m_links)
-    #endregion
-}
-
-/// <summary>
-/// code:MemorySampleSource hooks up a Memory graph to become a Sample source.  Currently we do
-/// a breadth-first traversal to form a spanning tree, and then create samples for each node
-/// where the 'stack' is the path to the root of this spanning tree.
-/// 
-/// This is just a first cut...
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class SpanningTree
-{
-    public SpanningTree(Graph graph, TextWriter log)
-    {
-        Graph = graph;
-        m_log = log;
-        m_nodeStorage = graph.AllocNodeStorage();
-        m_childStorage = graph.AllocNodeStorage();
-        m_typeStorage = graph.AllocTypeNodeStorage();
-
-        // We need to reduce the graph to a tree.   Each node is assigned a unique 'parent' which is its 
-        // parent in a spanning tree of the graph.  
-        // The +1 is for orphan node support.  
-        m_parent = new NodeIndex[(int)graph.NodeIndexLimit + 1];
-    }
-    public Graph Graph { get; }
-
-    /// <summary>
-    /// Every type is given a priority of 0 unless the type name matches one of 
-    /// the patterns in PriorityRegExs.  If it does that type is assigned that priority.
-    /// 
-    /// A node's priority is defined to be the priority of the type of the node
-    /// (as given by PriorityRegExs), plus 1/10 the priority of its parent.  
-    /// 
-    /// Thus priorities 'decay' by 1/10 through pointers IF the priority of the node's
-    /// type is 0 (the default).   
-    ///
-    /// By default the framework has a priority of -1 which means that you explore all
-    /// high priority and user defined types before any framework type.
-    /// 
-    /// Types with the same priority are enumerate breath-first.  
-    /// </summary>
-    public string PriorityRegExs
-    {
-        get
-        {
-            if (m_priorityRegExs == null)
-            {
-                PriorityRegExs = DefaultPriorities;
-            }
-
-            return m_priorityRegExs;
-        }
-        set
-        {
-            m_priorityRegExs = value;
-            SetTypePriorities(value);
-        }
-    }
-    public static string DefaultPriorities
-    {
-        get
-        {
-            return
-                // By types (including user defined types) are 0
-                @"v4.0.30319\%!->-1;" +     // Framework is less than default
-                @"v2.0.50727\%!->-1;" +     // Framework is less than default
-                @"[*local vars]->-1000;" +  // Local variables are not that interesting, since they tend to be transient
-                @"mscorlib!Runtime.CompilerServices.ConditionalWeakTable->-10000;" + // We prefer not to use Conditional weak table references even more. 
-                @"[COM/WinRT Objects]->-1000000;" + // We prefer to Not use the CCW roots. 
-                @"[*handles]->-2000000;" +
-                @"[other roots]->-2000000";
-        }
-    }
-
-    public NodeIndex Parent(NodeIndex node) { return m_parent[(int)node]; }
-
-    public void ForEach(Action<NodeIndex> callback)
-    {
-        // Initialize the priority 
-        if (m_typePriorities == null)
-        {
-            PriorityRegExs = DefaultPriorities;
-        }
-
-        Debug.Assert(m_typePriorities != null);
-
-        // Initialize the breadth-first work queue.
-        var nodesToVisit = new PriorityQueue(1024);
-        nodesToVisit.Enqueue(Graph.RootIndex, 0.0F);
-
-        // reset the visited information.
-        for (int i = 0; i < m_parent.Length; i++)
-        {
-            m_parent[i] = NodeIndex.Invalid;
-        }
-
-        float[] nodePriorities = new float[m_parent.Length];
-        bool scanedForOrphans = false;
-        var epsilon = 1E-7F;            // Something that is big enough not to bet lost in roundoff error.  
-        float order = 0;
-        for (int i = 0; ; i++)
-        {
-            if ((i & 0x1FFF) == 0)  // Every 8K
-            {
-                System.Threading.Thread.Sleep(0);       // Allow interruption.  
-            }
-
-            NodeIndex nodeIndex;
-            float nodePriority;
-            if (nodesToVisit.Count == 0)
-            {
-                nodePriority = 0;
-                if (!scanedForOrphans)
-                {
-                    scanedForOrphans = true;
-                    AddOrphansToQueue(nodesToVisit);
-                }
-                if (nodesToVisit.Count == 0)
-                {
-                    return;
-                }
-            }
-            nodeIndex = nodesToVisit.Dequeue(out nodePriority);
-
-            // Insert any children that have not already been visited (had a parent assigned) into the work queue). 
-            var node = Graph.GetNode(nodeIndex, m_nodeStorage);
-            var parentPriority = nodePriorities[(int)node.Index];
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                if (m_parent[(int)childIndex] == NodeIndex.Invalid)
-                {
-                    m_parent[(int)childIndex] = nodeIndex;
-
-                    // the priority of the child is determined by its type and 1/10 by its parent.  
-                    var child = Graph.GetNode(childIndex, m_childStorage);
-                    var childPriority = m_typePriorities[(int)child.TypeIndex] + parentPriority / 10;
-                    nodePriorities[(int)childIndex] = childPriority;
-
-                    // Subtract a small increasing value to keep the queue in order if the priorities are the same. 
-                    // This is a bit of a hack since it can get big and purtub the user-defined order.  
-                    order += epsilon;
-                    nodesToVisit.Enqueue(childIndex, childPriority - order);
-                }
-            }
-
-            // Return the node.  
-            callback?.Invoke(node.Index);
-        }
-    }
-
-    #region private
-    /// <summary>
-    /// Add any unreachable nodes to the 'nodesToVisit'.   Note that we do this in a 'smart' way
-    /// where we only add orphans that are not reachable from other orphans.   That way we get a 
-    /// minimal set of orphan 'roots'.  
-    /// </summary>
-    /// <param name="nodesToVisit"></param>
-    private void AddOrphansToQueue(PriorityQueue nodesToVisit)
-    {
-
-        for (int i = 0; i < (int)Graph.NodeIndexLimit; i++)
-        {
-            if (m_parent[i] == NodeIndex.Invalid)
-            {
-                MarkDecendentsIgnoringCycles((NodeIndex)i, 0);
-            }
-        }
-
-        // Collect up all the nodes that are not reachable from other nodes as the roots of the
-        // orphans.  Also reset orphanVisitedMarker back to NodeIndex.Invalid.
-        for (int i = 0; i < (int)Graph.NodeIndexLimit; i++)
-        {
-            var nodeIndex = (NodeIndex)i;
-            var parent = m_parent[(int)nodeIndex];
-            if (parent <= NodeIndex.Invalid)
-            {
-                if (parent == NodeIndex.Invalid)
-                {
-                    // Thr root index has no parent but is reachable from the root. 
-                    if (nodeIndex != Graph.RootIndex)
-                    {
-                        var node = Graph.GetNode(nodeIndex, m_nodeStorage);
-                        var priority = m_typePriorities[(int)node.TypeIndex];
-                        nodesToVisit.Enqueue(nodeIndex, priority);
-                        m_parent[(int)nodeIndex] = Graph.NodeIndexLimit;               // This is the 'not reachable' parent. 
-                    }
-                }
-                else
-                {
-                    m_parent[(int)nodeIndex] = NodeIndex.Invalid;
-                }
-            }
-        }
-    }
-
-    /// <summary>
-    /// A helper for AddOrphansToQueue, so we only add orphans that are not reachable from other orphans.  
-    /// 
-    /// Mark all descendants (but not nodeIndex itself) as being visited.    Any arcs that form
-    /// cycles are ignored, so nodeIndex is guaranteed to NOT be marked.     
-    /// </summary>
-    private void MarkDecendentsIgnoringCycles(NodeIndex nodeIndex, int recursionCount)
-    {
-        // TODO We give up if the chains are larger than 10K long (because we stack overflow otherwise)
-        // We could have an explicit stack and avoid this...
-        if (recursionCount > 10000)
-        {
-            return;
-        }
-
-        Debug.Assert(m_parent[(int)nodeIndex] == NodeIndex.Invalid);
-
-        // This marks that there is a path from another orphan to this one (thus it is not a good root)
-        NodeIndex orphanVisitedMarker = NodeIndex.Invalid - 1;
-
-        // To detect cycles we mark all nodes we not committed to (we are visiting, rather than visited)
-        // If we detect this mark we understand it is a loop and ignore the arc.  
-        NodeIndex orphanVisitingMarker = NodeIndex.Invalid - 2;
-        m_parent[(int)nodeIndex] = orphanVisitingMarker;        // We are now visitING
-
-        // Mark all nodes as being visited.  
-        var node = Graph.GetNode(nodeIndex, AllocNodeStorage());
-        for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-        {
-            // Has this child not been seen at all?  If so mark it.  
-            // Skip it if we are visiting (it would form a cycle) or visited (or not an orphan)
-            if (m_parent[(int)childIndex] == NodeIndex.Invalid)
-            {
-                MarkDecendentsIgnoringCycles(childIndex, recursionCount + 1);
-                m_parent[(int)childIndex] = orphanVisitedMarker;
-            }
-        }
-        FreeNodeStorage(node);
-
-        // We set this above, and should not have changed it.  
-        Debug.Assert(m_parent[(int)nodeIndex] == orphanVisitingMarker);
-        // Now that we are finished, we reset the visiting bit.  
-        m_parent[(int)nodeIndex] = NodeIndex.Invalid;
-    }
-
-    /// <summary>
-    /// Gives back nodes that are no longer in use.  This is a memory optimization. 
-    /// </summary>
-    private void FreeNodeStorage(Node node)
-    {
-        m_cachedNodeStorage = node;
-    }
-    /// <summary>
-    /// Gets a node that can be written on.  It is a simple cache
-    /// </summary>
-    /// <returns></returns>
-    private Node AllocNodeStorage()
-    {
-        var ret = m_cachedNodeStorage;                // See if we have a free node. 
-        if (ret == null)
-        {
-            ret = Graph.AllocNodeStorage();
-        }
-        else
-        {
-            m_cachedNodeStorage = null;               // mark that that node is in use.  
-        }
-
-        return ret;
-    }
-
-    /// <summary>
-    /// Convert a string from my regular expression format (where you only have * and {  } as grouping operators
-    /// and convert them to .NET regular expressions string
-    /// TODO FIX NOW cloned code (also in FilterStackSource)
-    /// </summary>
-    internal static string ToDotNetRegEx(string str)
-    {
-        // A leading @ sign means the rest is a .NET regular expression.  (Undocumented, not really needed yet.)
-        if (str.StartsWith('@'))
-        {
-            return str.Substring(1);
-        }
-
-        str = Regex.Escape(str);                // Assume everything is ordinary
-        str = str.Replace(@"%", @"[.\w\d?]*");  // % means any number of alpha-numeric chars. 
-        str = str.Replace(@"\*", @".*");        // * means any number of any characters.  
-        str = str.Replace(@"\^", @"^");         // ^ means anchor at the beginning.  
-        str = str.Replace(@"\|", @"|");         // | means is the or operator  
-        str = str.Replace(@"\{", "(");
-        str = str.Replace('}', ')');
-        return str;
-    }
-
-    private void SetTypePriorities(string priorityPats)
-    {
-        m_typePriorities ??= new float[(int)Graph.NodeTypeIndexLimit];
-
-        string[] priorityPatArray = priorityPats.Split(';');
-        Regex[] priorityRegExArray = new Regex[priorityPatArray.Length];
-        float[] priorityArray = new float[priorityPatArray.Length];
-        for (int i = 0; i < priorityPatArray.Length; i++)
-        {
-            var m = Regex.Match(priorityPatArray[i], @"(.*)->(-?\d+.?\d*)");
-            if (!m.Success)
-            {
-                if (string.IsNullOrWhiteSpace(priorityPatArray[i]))
-                {
-                    continue;
-                }
-
-                throw new ApplicationException($"Priority pattern {priorityPatArray[i]} is not of the form Pat->Num.");
-            }
-
-            var dotNetRegEx = ToDotNetRegEx(m.Groups[1].Value.Trim());
-            priorityRegExArray[i] = new Regex(dotNetRegEx, RegexOptions.IgnoreCase);
-            priorityArray[i] = float.Parse(m.Groups[2].Value);
-        }
-
-        // Assign every type index a priority in m_typePriorities based on if they match a pattern.  
-        NodeType typeStorage = Graph.AllocTypeNodeStorage();
-        for (NodeTypeIndex typeIdx = 0; typeIdx < Graph.NodeTypeIndexLimit; typeIdx++)
-        {
-            var type = Graph.GetType(typeIdx, typeStorage);
-
-            var fullName = type.FullName;
-            for (int regExIdx = 0; regExIdx < priorityRegExArray.Length; regExIdx++)
-            {
-                var priorityRegEx = priorityRegExArray[regExIdx];
-                if (priorityRegEx == null)
-                {
-                    continue;
-                }
-
-                var m = priorityRegEx.Match(fullName);
-                if (m.Success)
-                {
-                    m_typePriorities[(int)typeIdx] = priorityArray[regExIdx];
-                    // m_log.WriteLine("Type {0} assigned priority {1:f3}", fullName, priorityArray[regExIdx]);
-                    break;
-                }
-            }
-        }
-    }
-
-    private NodeIndex[] m_parent;               // We keep track of the parents of each node in our breadth-first scan. 
-
-    // We give each type a priority (using the m_priority Regular expressions) which guide the breadth-first scan. 
-    private string m_priorityRegExs;
-    private float[] m_typePriorities;
-    private NodeType m_typeStorage;
-    private Node m_nodeStorage;                 // Only for things that can't be reentrant
-    private Node m_childStorage;
-    private Node m_cachedNodeStorage;           // Used when it could be reentrant
-    private TextWriter m_log;                   // processing messages 
-    #endregion
-}
-
-/// <summary>
-/// TODO FIX NOW put in its own file.  
-/// A priority queue, specialized to be a bit more efficient than a generic version would be. 
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class PriorityQueue
-{
-    public PriorityQueue(int initialSize = 32)
-    {
-        m_heap = new DataItem[initialSize];
-    }
-    public int Count { get; private set; }
-
-    public void Enqueue(NodeIndex item, float priority)
-    {
-        var idx = Count;
-        if (idx >= m_heap.Length)
-        {
-            var newArray = new DataItem[m_heap.Length * 3 / 2 + 8];
-            Array.Copy(m_heap, newArray, m_heap.Length);
-            m_heap = newArray;
-        }
-        m_heap[idx].value = item;
-        m_heap[idx].priority = priority;
-        Count = idx + 1;
-        for (; ; )
-        {
-            var parent = idx / 2;
-            if (m_heap[parent].priority >= m_heap[idx].priority)
-            {
-                break;
-            }
-
-            (m_heap[idx], m_heap[parent]) = (m_heap[parent], m_heap[idx]);
-
-            if (parent == 0)
-            {
-                break;
-            }
-
-            idx = parent;
-        }
-        // CheckInvariant();
-    }
-    public NodeIndex Dequeue(out float priority)
-    {
-        Debug.Assert(Count > 0);
-
-        var ret = m_heap[0].value;
-        priority = m_heap[0].priority;
-        --Count;
-        m_heap[0] = m_heap[Count];
-        var idx = 0;
-        for (; ; )
-        {
-            var childIdx = idx * 2;
-            var largestIdx = idx;
-            if (childIdx < Count && m_heap[childIdx].priority > m_heap[largestIdx].priority)
-            {
-                largestIdx = childIdx;
-            }
-
-            childIdx++;
-            if (childIdx < Count && m_heap[childIdx].priority > m_heap[largestIdx].priority)
-            {
-                largestIdx = childIdx;
-            }
-
-            if (largestIdx == idx)
-            {
-                break;
-            }
-
-            (m_heap[idx], m_heap[largestIdx]) = (m_heap[largestIdx], m_heap[idx]);
-
-            idx = largestIdx;
-        }
-        // CheckInvariant();
-        return ret;
-    }
-
-    #region private
-#if DEBUG
-    public override string ToString()
-    {
-        var sb = new StringBuilder();
-        sb.AppendLine("<PriorityQueue Count=\"").Append(Count).Append("\">").AppendLine();
-
-        // Sort the items in descending order 
-        var items = new List<DataItem>(Count);
-        for (int i = 0; i < Count; i++)
-            items.Add(m_heap[i]);
-        items.Sort((x, y) => y.priority.CompareTo(x.priority));
-        if (items.Count > 0)
-            Debug.Assert(items[0].value == m_heap[0].value);
-
-        foreach (var item in items)
-            sb.Append("{").Append((int)item.value).Append(", ").Append(item.priority.ToString("f1")).Append("}").AppendLine();
-        sb.AppendLine("</PriorityQueue>");
-        return sb.ToString();
-    }
-#endif
-
-    private struct DataItem
-    {
-        public DataItem(NodeIndex value, float priority) { this.value = value; this.priority = priority; }
-        public float priority;
-        public NodeIndex value;
-    }
-    [Conditional("DEBUG")]
-    private void CheckInvariant()
-    {
-        for (int idx = 1; idx < Count; idx++)
-        {
-            var parentIdx = idx / 2;
-            Debug.Assert(m_heap[parentIdx].priority >= m_heap[idx].priority);
-        }
-    }
-
-    // In this array form a tree where each child of i is at 2i and 2i+1.   Each child is 
-    // less than or equal to its parent.  
-    private DataItem[] m_heap;
-
-    #endregion
-}
-
-/// <summary>
-/// This class is responsible for taking a graph and generating a smaller graph that
-/// is a reasonable proxy.   In particular
-///     
-///     1) A spanning tree is formed, and if a node is selected so are all its 
-///        parents in that spanning tree.
-///        
-///     2) We try hard to keep scale each object type by the count by which the whole
-///        graph was reduced.  
-/// </summary>
-[ExcludeFromCodeCoverage]
-internal sealed class GraphSampler
-{
-    /// <summary>
-    /// 
-    /// </summary>
-    public GraphSampler(MemoryGraph graph, int targetNodeCount, TextWriter log)
-    {
-        m_graph = graph;
-        m_log = log;
-        m_targetNodeCount = targetNodeCount;
-        m_filteringRatio = (float)graph.NodeCount / targetNodeCount;
-        m_nodeStorage = m_graph.AllocNodeStorage();
-        m_childNodeStorage = m_graph.AllocNodeStorage();
-        m_nodeTypeStorage = m_graph.AllocTypeNodeStorage();
-    }
-
-    /// <summary>
-    /// Creates a new graph from 'graph' which has the same type statistics as the original
-    /// graph but keeps the node count roughly at 'targetNodeCount'
-    /// </summary>
-    public MemoryGraph GetSampledGraph()
-    {
-        m_log.WriteLine("************* SAMPLING GRAPH TO REDUCE SIZE ***************");
-        m_log.WriteLine("Original graph object count {0:n0}, targetObjectCount {1:n0} targetRatio {2:f2}", m_graph.NodeCount, m_targetNodeCount, m_filteringRatio);
-        m_log.WriteLine("Original graph Size MB {0:n0} TypeCount {1:n0}", m_graph.TotalSize, m_graph.NodeTypeCount);
-
-        // Get the spanning tree
-        m_spanningTree = new SpanningTree(m_graph, m_log);
-        m_spanningTree.ForEach(null);
-
-        // Make the new graph 
-        m_newGraph = new MemoryGraph(m_targetNodeCount + m_graph.NodeTypeCount * 2);
-        m_newGraph.Is64Bit = m_graph.Is64Bit;
-
-        // Initialize the object statistics
-        m_statsByType = new SampleStats[m_graph.NodeTypeCount];
-        for (int i = 0; i < m_statsByType.Length; i++)
-        {
-            m_statsByType[i] = new SampleStats();
-        }
-
-        // And initialize the mapping from old nodes to new nodes.  (TODO: this can be a hash table to save size?  )
-        m_newIndex = new NodeIndex[m_graph.NodeCount];
-        for (int i = 0; i < m_newIndex.Length; i++)
-        {
-            m_newIndex[i] = NodeIndex.Invalid;
-        }
-
-        ValidateStats(false);
-
-        VisitNode(m_graph.RootIndex, true, false); // visit the root for sure.  
-        // Sample the nodes, trying to keep the 
-        for (NodeIndex nodeIdx = 0; nodeIdx < m_graph.NodeIndexLimit; nodeIdx++)
-        {
-            VisitNode(nodeIdx, false, false);
-        }
-
-        ValidateStats(true);
-
-        // See if we need to flesh out the potential node to become truly sampled node to hit our quota.  
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            var newIndex = m_newIndex[(int)nodeIdx];
-            if (newIndex == PotentialNode)
-            {
-                var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-                var stats = m_statsByType[(int)node.TypeIndex];
-                int quota = (int)(stats.TotalCount / m_filteringRatio + .5);
-                int needed = quota - stats.SampleCount;
-                if (needed > 0)
-                {
-                    // If we have not computed the frequency of sampling do it now.  
-                    if (stats.SkipFreq == 0)
-                    {
-                        var available = stats.PotentialCount - stats.SampleCount;
-                        Debug.Assert(0 <= available);
-                        Debug.Assert(needed <= available);
-                        stats.SkipFreq = Math.Max(available / needed, 1);
-                    }
-
-                    // Sample every Nth time.  
-                    stats.SkipCtr++;
-                    if (stats.SkipFreq <= stats.SkipCtr)
-                    {
-                        // Sample a new node 
-                        m_newIndex[(int)nodeIdx] = m_newGraph.CreateNode();
-                        stats.SampleCount++;
-                        stats.SampleMetric += node.Size;
-                        stats.SkipCtr = 0;
-                    }
-                }
-            }
-        }
-
-        // OK now m_newIndex tell us which nodes we want.  actually define the selected nodes. 
-
-        // Initialize the mapping from old types to new types.
-        m_newTypeIndexes = new NodeTypeIndex[m_graph.NodeTypeCount];
-        for (int i = 0; i < m_newTypeIndexes.Length; i++)
-        {
-            m_newTypeIndexes[i] = NodeTypeIndex.Invalid;
-        }
-
-        GrowableArray<NodeIndex> children = new GrowableArray<NodeIndex>(100);
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            // Add all sampled nodes to the new graph.  
-            var newIndex = m_newIndex[(int)nodeIdx];
-            if (IsSampledNode(newIndex))
-            {
-                var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-                // Get the children that are part of the sample (ignore ones that are filter)
-                children.Clear();
-                for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-                {
-                    var newChildIndex = m_newIndex[(int)childIndex];
-                    if (0 <= newChildIndex)                 // the child is not filtered out. 
-                    {
-                        children.Add(newChildIndex);
-                    }
-                }
-                // define the node
-                var newTypeIndex = GetNewTypeIndex(node.TypeIndex);
-                m_newGraph.SetNode(newIndex, newTypeIndex, node.Size, children);
-                m_newGraph.SetAddress(newIndex, m_graph.GetAddress(nodeIdx));
-            }
-        }
-
-        ValidateStats(true, true);
-
-        // Set the root.
-        m_newGraph.RootIndex = m_newIndex[(int)m_graph.RootIndex];
-        Debug.Assert(0 <= m_newGraph.RootIndex);            // RootIndex in the tree.  
-
-        m_newGraph.AllowReading();
-
-        // At this point we are done.  The rest is just to report the result to the user.  
-
-        // Sort the m_statsByType
-        var sortedTypes = new int[m_statsByType.Length];
-        for (int i = 0; i < sortedTypes.Length; i++)
-        {
-            sortedTypes[i] = i;
-        }
-
-        Array.Sort(sortedTypes, delegate (int x, int y)
-        {
-            var ret = m_statsByType[y].TotalMetric.CompareTo(m_statsByType[x].TotalMetric);
-            return ret;
-        });
-
-        m_log.WriteLine("Stats of the top types (out of {0:n0})", m_newGraph.NodeTypeCount);
-        m_log.WriteLine("OrigSizeMeg SampleSizeMeg   Ratio   |   OrigCnt  SampleCnt    Ratio   | Ave Size | Type Name");
-        m_log.WriteLine("---------------------------------------------------------------------------------------------");
-
-        for (int i = 0; i < Math.Min(m_statsByType.Length, 30); i++)
-        {
-            int typeIdx = sortedTypes[i];
-            NodeType type = m_graph.GetType((NodeTypeIndex)typeIdx, m_nodeTypeStorage);
-            var stats = m_statsByType[typeIdx];
-
-            m_log.WriteLine("{0,12:n6} {1,11:n6}  {2,9:f2} | {3,10:n0} {4,9:n0}  {5,9:f2} | {6,8:f0} | {7}",
-                stats.TotalMetric / 1000000.0, stats.SampleMetric / 1000000.0, stats.SampleMetric == 0 ? 0.0 : (double)stats.TotalMetric / stats.SampleMetric,
-                stats.TotalCount, stats.SampleCount, stats.SampleCount == 0 ? 0.0 : (double)stats.TotalCount / stats.SampleCount,
-                (double)stats.TotalMetric / stats.TotalCount, type.Name);
-        }
-
-        m_log.WriteLine("Sampled Graph node count {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.NodeCount,
-            (double)m_graph.NodeCount / m_newGraph.NodeCount);
-        m_log.WriteLine("Sampled Graph type count {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.NodeTypeCount,
-            (double)m_graph.NodeTypeCount / m_newGraph.NodeTypeCount);
-        m_log.WriteLine("Sampled Graph node size  {0,11:n0} (reduced by {1:f2} ratio)", m_newGraph.TotalSize,
-            (double)m_graph.TotalSize / m_newGraph.TotalSize);
-        return m_newGraph;
-    }
-
-    /// <summary>
-    /// returns an array of scaling factors.  This array is indexed by the type index of
-    /// the returned graph returned by GetSampledGraph.   If the sampled count for that type multiplied
-    /// by this scaling factor, you end up with the count for that type of the original unsampled graph.  
-    /// </summary>
-    public float[] CountScalingByType
-    {
-        get
-        {
-            var ret = new float[m_newGraph.NodeTypeCount];
-            for (int i = 0; i < m_statsByType.Length; i++)
-            {
-                var newTypeIndex = MapTypeIndex((NodeTypeIndex)i);
-                if (newTypeIndex != NodeTypeIndex.Invalid)
-                {
-                    float scale = 1;
-                    if (m_statsByType[i].SampleMetric != 0)
-                    {
-                        scale = (float)((double)m_statsByType[i].TotalMetric / m_statsByType[i].SampleMetric);
-                    }
-
-                    ret[(int)newTypeIndex] = scale;
-                }
-            }
-            for (int i = 1; i < ret.Length; i++)
-            {
-                Debug.Assert(0 < ret[i] && ret[i] <= float.MaxValue);
-            }
-
-            return ret;
-        }
-    }
-
-    /// <summary>
-    /// Maps 'oldTypeIndex' to its type index in the output graph
-    /// </summary>
-    /// <returns>New type index, will be Invalid if the type is not in the output graph</returns>
-    public NodeTypeIndex MapTypeIndex(NodeTypeIndex oldTypeIndex)
-    {
-        return m_newTypeIndexes[(int)oldTypeIndex];
-    }
-
-    /// <summary>
-    /// Maps 'oldNodeIndex' to its new node index in the output graph
-    /// </summary>
-    /// <returns>New node index, will be less than 0 if the node is not in the output graph</returns>
-    public NodeIndex MapNodeIndex(NodeIndex oldNodeIndex)
-    {
-        return m_newIndex[(int)oldNodeIndex];
-    }
-
-    #region private
-    /// <summary>
-    /// Visits 'nodeIdx', if already visited, do nothing.  If unvisited determine if 
-    /// you should add this node to the graph being built.   If 'mustAdd' is true or
-    /// if we need samples it keep the right sample/total ratio, then add the sample.  
-    /// </summary>
-    private void VisitNode(NodeIndex nodeIdx, bool mustAdd, bool doNotAddAncestors)
-    {
-        var newNodeIdx = m_newIndex[(int)nodeIdx];
-        // If this node has been selected already, there is nothing to do.    
-        if (IsSampledNode(newNodeIdx))
-        {
-            return;
-        }
-        // If we have visited this node and reject it and we are not forced to add it, we are done.
-        if (newNodeIdx == RejectedNode && !mustAdd)
-        {
-            return;
-        }
-
-        Debug.Assert(newNodeIdx == NodeIndex.Invalid || newNodeIdx == PotentialNode || (newNodeIdx == RejectedNode && mustAdd));
-
-        var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-        var stats = m_statsByType[(int)node.TypeIndex];
-
-        // If we have never seen this node before, add to our total count.  
-        if (newNodeIdx == NodeIndex.Invalid)
-        {
-            if (stats.TotalCount == 0)
-            {
-                m_numDistictTypes++;
-            }
-
-            stats.TotalCount++;
-            stats.TotalMetric += node.Size;
-        }
-
-        // Also insure that if there are a large number of types, that we sample them at least some. 
-        if (stats.SampleCount == 0 && !mustAdd && (m_numDistictTypesWithSamples + .5F) * m_filteringRatio <= m_numDistictTypes)
-        {
-            mustAdd = true;
-        }
-
-        // We sample if we are forced (it is part of a parent chain), we need it to 
-        // mimic the the original statistics, or if it is a large object (we include 
-        // all large objects, since the affect overall stats so much).  
-        if (mustAdd ||
-            (stats.PotentialCount + .5f) * m_filteringRatio <= stats.TotalCount ||
-            85000 < node.Size)
-        {
-            if (stats.SampleCount == 0)
-            {
-                m_numDistictTypesWithSamples++;
-            }
-
-            stats.SampleCount++;
-            stats.SampleMetric += node.Size;
-            if (newNodeIdx != PotentialNode)
-            {
-                stats.PotentialCount++;
-            }
-
-            m_newIndex[(int)nodeIdx] = m_newGraph.CreateNode();
-
-            // Add all direct children as potential nodes (Potential nodes I can add without adding any other node)
-            for (var childIndex = node.GetFirstChildIndex(); childIndex != NodeIndex.Invalid; childIndex = node.GetNextChildIndex())
-            {
-                var newChildIndex = m_newIndex[(int)childIndex];
-                // Already a sampled or potential node.  Nothing to do.  
-                if (IsSampledNode(newChildIndex) || newChildIndex == PotentialNode)
-                {
-                    continue;
-                }
-
-                var childNode = m_graph.GetNode(childIndex, m_childNodeStorage);
-                var childStats = m_statsByType[(int)childNode.TypeIndex];
-
-                if (newChildIndex == NodeIndex.Invalid)
-                {
-                    if (stats.TotalCount == 0)
-                    {
-                        m_numDistictTypes++;
-                    }
-
-                    childStats.TotalCount++;
-                    childStats.TotalMetric += childNode.Size;
-                }
-                else
-                {
-                    Debug.Assert(newChildIndex == RejectedNode);
-                }
-
-                m_newIndex[(int)childIndex] = PotentialNode;
-                childStats.PotentialCount++;
-            }
-
-            // For all ancestors, require them to be in the list
-            if (!doNotAddAncestors)
-            {
-                for (; ; )
-                {
-                    nodeIdx = m_spanningTree.Parent(nodeIdx);
-                    if (nodeIdx == NodeIndex.Invalid || m_newIndex.Length == (int)nodeIdx) // The last index represents the 'orphan' node.  
-                    {
-                        break;
-                    }
-
-                    // Indicate that you should not add ancestors (since I will do this).  
-                    // We do the adding in a loop (rather than letting recursion do it) to avoid stack overflows
-                    // for long chains of objects.  
-                    VisitNode(nodeIdx, true, true);
-                }
-            }
-        }
-        else
-        {
-            if (newNodeIdx != PotentialNode)
-            {
-                m_newIndex[(int)nodeIdx] = RejectedNode;
-            }
-        }
-    }
-
-    /// <summary>
-    /// Maps 'oldTypeIndex' to its type index in the output graph
-    /// </summary>
-    /// <param name="oldTypeIndex"></param>
-    /// <returns></returns>
-    private NodeTypeIndex GetNewTypeIndex(NodeTypeIndex oldTypeIndex)
-    {
-        var ret = m_newTypeIndexes[(int)oldTypeIndex];
-        if (ret == NodeTypeIndex.Invalid)
-        {
-            var oldType = m_graph.GetType(oldTypeIndex, m_nodeTypeStorage);
-            ret = m_newGraph.CreateType(oldType.Name, oldType.ModuleName, oldType.Size);
-            m_newTypeIndexes[(int)oldTypeIndex] = ret;
-        }
-        return ret;
-    }
-
-
-    [Conditional("DEBUG")]
-    private void ValidateStats(bool allNodesVisited, bool completed = false)
-    {
-        var statsCheckByType = new SampleStats[m_statsByType.Length];
-        for (int i = 0; i < statsCheckByType.Length; i++)
-        {
-            statsCheckByType[i] = new SampleStats();
-        }
-
-        int total = 0;
-        long totalSize = 0;
-        int sampleTotal = 0;
-        var typeStorage = m_graph.AllocTypeNodeStorage();
-        for (NodeIndex nodeIdx = 0; nodeIdx < (NodeIndex)m_newIndex.Length; nodeIdx++)
-        {
-            var node = m_graph.GetNode(nodeIdx, m_nodeStorage);
-            var stats = statsCheckByType[(int)node.TypeIndex];
-            var type = node.GetType(typeStorage);
-            var typeName = type.Name;
-            var newNodeIdx = m_newIndex[(int)nodeIdx];
-
-            if (newNodeIdx == NodeIndex.Invalid)
-            {
-                // We should have visited every node, so there should be no Invalid nodes. 
-                Debug.Assert(!allNodesVisited);
-            }
-            else
-            {
-                total++;
-                stats.TotalCount++;
-                stats.TotalMetric += node.Size;
-                totalSize += node.Size;
-                Debug.Assert(node.Size != 0 || typeName.StartsWith('[') || typeName == "UNDEFINED");
-                if (IsSampledNode(newNodeIdx) || newNodeIdx == PotentialNode)
-                {
-                    if (nodeIdx != m_graph.RootIndex)
-                    {
-                        Debug.Assert(IsSampledNode(m_spanningTree.Parent(nodeIdx)));
-                    }
-
-                    stats.PotentialCount++;
-                    if (IsSampledNode(newNodeIdx))
-                    {
-                        stats.SampleCount++;
-                        sampleTotal++;
-                        stats.SampleMetric += node.Size;
-                    }
-                }
-                else
-                {
-                    Debug.Assert(newNodeIdx == RejectedNode);
-                }
-            }
-            statsCheckByType[(int)node.TypeIndex] = stats;
-        }
-
-        float[] scalings = null;
-        if (completed)
-        {
-            scalings = CountScalingByType;
-        }
-
-        for (NodeTypeIndex typeIdx = 0; typeIdx < m_graph.NodeTypeIndexLimit; typeIdx++)
-        {
-            _ = m_graph.GetType(typeIdx, typeStorage);
-            var statsCheck = statsCheckByType[(int)typeIdx];
-            var stats = m_statsByType[(int)typeIdx];
-
-            Debug.Assert(stats.TotalMetric == statsCheck.TotalMetric);
-            Debug.Assert(stats.TotalCount == statsCheck.TotalCount);
-            Debug.Assert(stats.SampleCount == statsCheck.SampleCount);
-            Debug.Assert(stats.SampleMetric == statsCheck.SampleMetric);
-            Debug.Assert(stats.PotentialCount == statsCheck.PotentialCount);
-
-            Debug.Assert(stats.PotentialCount <= statsCheck.TotalCount);
-            Debug.Assert(stats.SampleCount <= statsCheck.PotentialCount);
-
-            // We should be have at least m_filterRatio of Potential objects 
-            Debug.Assert(!((stats.PotentialCount + .5f) * m_filteringRatio <= stats.TotalCount));
-
-            // If we completed, then we converted potentials to true samples.   
-            if (completed)
-            {
-                Debug.Assert(!((stats.SampleCount + .5f) * m_filteringRatio <= stats.TotalCount));
-
-                // Make sure that scalings that we finally output were created correctly
-                if (stats.SampleMetric > 0)
-                {
-                    var newTypeIdx = MapTypeIndex(typeIdx);
-                    var estimatedTotalMetric = scalings[(int)newTypeIdx] * stats.SampleMetric;
-                    Debug.Assert(Math.Abs(estimatedTotalMetric - stats.TotalMetric) / stats.TotalMetric < .01);
-                }
-            }
-
-            if (stats.SampleCount == 0)
-            {
-                Debug.Assert(stats.SampleMetric == 0);
-            }
-
-            if (stats.TotalMetric == 0)
-            {
-                Debug.Assert(stats.TotalMetric == 0);
-            }
-        }
-
-        if (allNodesVisited)
-        {
-            Debug.Assert(total == m_graph.NodeCount);
-            // TODO The assert should be Debug.Assert(totalSize == m_graph.TotalSize);
-            // but we have to give a 1% error margin to get things passing. Fix this.
-            Debug.Assert(Math.Abs(totalSize - m_graph.TotalSize) / totalSize < .01);
-        }
-        Debug.Assert(sampleTotal == m_newGraph.NodeCount);
-    }
-
-    private sealed class SampleStats
-    {
-        public int TotalCount;          // The number of objects of this type in the original graph
-        public int SampleCount;         // The number of objects of this type we have currently added to the new graph
-        public int PotentialCount;      // SampleCount + The number of objects of this type that can be added without needing to add other nodes
-        public long TotalMetric;
-        public long SampleMetric;
-        public int SkipFreq;          // When sampling potentials, take every Nth one where this is the N
-        public int SkipCtr;             // This remembers our last N.  
-    };
-
-    /// <summary>
-    /// This value goes in the m_newIndex[].   If we accept the node into the sampled graph, we put the node
-    /// index in the NET graph in m_newIndex.   If we reject the node we use the special RejectedNode value
-    /// below
-    /// </summary>
-    private const NodeIndex RejectedNode = (NodeIndex)(-2);
-
-    /// <summary>
-    /// This value also goes in m_newIndex[].   If we can add this node without needing to add any other nodes
-    /// to the new graph (that is it is one hop from an existing accepted node, then we mark it specially as
-    /// a PotentialNode).   We add these in a second pass over the data.  
-    /// </summary>
-    private const NodeIndex PotentialNode = (NodeIndex)(-3);
-
-    private bool IsSampledNode(NodeIndex nodeIdx) { return 0 <= nodeIdx; }
-
-    private MemoryGraph m_graph;
-    private int m_targetNodeCount;
-    private TextWriter m_log;
-    private Node m_nodeStorage;
-    private Node m_childNodeStorage;
-    private NodeType m_nodeTypeStorage;
-    private float m_filteringRatio;
-    private SampleStats[] m_statsByType;
-    private int m_numDistictTypesWithSamples;
-    private int m_numDistictTypes;
-    private NodeIndex[] m_newIndex;
-    private NodeTypeIndex[] m_newTypeIndexes;
-    private SpanningTree m_spanningTree;
-    private MemoryGraph m_newGraph;
-    #endregion
 }
 
 #if false
